+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Changes:
 * Remove `--shell-version` and `--terminal-version`. They are always enabled
 * Remove `--*-error-format`, which seems to be useless
 * Remove `--display-detect-name`. Display name is always detected, and will be printed if multiple displays are detected
-* Remove `--set` and `--set-keyless`. Use JSON config with Custom module instead
+* Deprecate `--set` and `--set-keyless`; they may be removed in future releases. Use JSON config with Custom module instead
+* Remove the special handling of Command module (it can be set once in the triditional `config.conf`). Use JSON config with Command module instead
 
 Features:
 * FreeBSD support is improved greatly, and actually tested in a phycial machine